backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
neo4j==5.12.0
pandas==2.1.1
pydantic==2.4.2